+++ packages/core/src/transport.ts
@@ -8,6 +8,8 @@ export interface SendOptions {
   method: string
   /** Request HTTP headers. */
   headers?: {[key: string]: string}
+  /** When specified, message body larger than the treshold is gzipped  */
+  gzipThreshold?: number
 }
 
 /**